Revert "x86/acpi: Move acpi_wakeup_cpu() and helpers to smpwakeup.c"
This reverts commit 672fc21ee1aef5e73a8ba630d676d6e1d966a889.
